Flight detail view implemented

+++ app/src/main/java/com/example/recruitmenttask/model/Fare.kt
@@ -1,5 +1,7 @@
 package com.example.recruitmenttask.model
 
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
 import java.io.Serializable
 
 data class Fare(
@@ -8,7 +10,7 @@ data class Fare(
     val count: Int,
     val hasDiscount: Boolean,
     val publishedFare: Double,
-    val discountInPercent: Double,
+    val discountInPercent: Int,
     val hasPromoDiscount: Boolean,
     val discountAmount: Double,
     val hasBogof: Boolean
